+++ internal/cli/src/utils.ts
@@ -1,0 +1,14 @@
+import assert from 'assert'
+import escaladeSync from 'escalade/sync'
+import path from 'path'
+
+// eslint-disable-next-line consistent-return
+const pkgsDir = escaladeSync(__dirname, (_dir, names) => {
+  if (names.includes('packages')) {
+    return 'packages'
+  }
+})
+assert(pkgsDir, `can not find "packages" dir from ${__dirname}`)
+
+export const pkgsDirJoin = (...args: string[]): string =>
+  path.join(pkgsDir, ...args)