+++ packages/cordova/ts/admob.ts
@@ -6,48 +6,7 @@ import { execAsync, fireDocumentEvent, NativeActions } from './base'
 import Interstitial from './interstitial'
 import RewardVideo from './reward-video'
 import AdMobState from './state'
-
-type MobileAdOptions = { adUnitId: string }
-
-class MobileAd {
-  private static allAds: { [s: number]: MobileAd } = {}
-
-  private _id: number
-  public adUnitId: string
-
-  constructor({ adUnitId }: MobileAdOptions) {
-    this.adUnitId = adUnitId
-
-    this._id = 10001 + Object.keys(MobileAd.allAds).length
-    MobileAd.allAds[this.id] = this
-  }
-
-  public get id() {
-    return this._id
-  }
-}
-
-class BannerAd extends MobileAd {
-  constructor({ adUnitId }: MobileAdOptions) {
-    super({ adUnitId })
-  }
-
-  public show(opts: IBannerRequest) {
-    return execAsync(NativeActions.banner_show, [
-      {
-        position: 'bottom',
-        size: AdSizeType.SMART_BANNER,
-        ...opts,
-        adUnitID: this.adUnitId,
-        id: this.id,
-      },
-    ])
-  }
-
-  public hide() {
-    return execAsync(NativeActions.banner_hide, [{ id: this.id }])
-  }
-}
+import { BannerAd } from './banner'
 
 class AdMob {
   public banner: Banner