+++ scripts/utils.ts
@@ -1,4 +1,0 @@
-import * as path from 'path'
-
-export const pkgsDirJoin = (...args: string[]) =>
-  path.join(__dirname, '../packages', ...args)