+++ packages/cordova/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.40.0](https://github.com/admob-plus/admob-plus/compare/cordova-admob-plus@0.39.0...cordova-admob-plus@0.40.0) (2020-11-22)
+
+
+### Features
+
+* **cordova:** update default PLAY_SERVICES_VERSION to 19.5.0 ([ee541ff](https://github.com/admob-plus/admob-plus/commit/ee541ff753405bd526a778e93ede8ae179c8cd8a))
+* **cordova:** update Google-Mobile-Ads-SDK to v7.68 ([49ae810](https://github.com/admob-plus/admob-plus/commit/49ae810c9fd80364a0a34decd4e0c1921a926133))
+
+
+
+
+
 # [0.39.0](https://github.com/admob-plus/admob-plus/compare/cordova-admob-plus@0.37.0...cordova-admob-plus@0.39.0) (2020-10-18)